backend/requirements.txt
streamlit
pandas
dnspython
pypdf
langchain
langchain-community
langchain-openai
langchain-groq
python-dotenv
flask
flask-cors
